JBIDE-16140 buildchow for templated jenkins jobs
Initial contribution that provide setup to use jenkins-job-builder
to generate jenkins jobs.

Tried to keep it simple but yet structured to avoid
having a very large hard to maintain single yaml file.

Detailed docs are in the readme.adoc, but the general gist is:

templates - contains job templates.
inis - contain setup for specific hosts `defaults` being default jenkins.
       dockerhost folder works as example to see how to do overrides.

The main template is `component-template.yaml` which defines the component jobs.

To generate jobs just run:

   cd inis/defaults
   jenkins-job-builder --config jenkins_jobs.ini test -o generated

Note: for now the component jobs should match what was in place 2015-06-01
in the original config.xml, but the rest have not been verified yet.